# swagger.json
{"openapi":"3.0.0","paths":{"/session/login":{"post":{"operationId":"SessionController_postLogin","summary":"Post the login credentials data","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostLoginReqDto"}}}},"responses":{"200":{"description":"Returns a JSON with the jwt data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostLoginResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["session"]}},"/user/get":{"get":{"operationId":"UserController_getUser","summary":"Get the user data","parameters":[],"responses":{"200":{"description":"Returns a JSON with the user data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetUserResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["user"],"security":[{"auth":[]}]}},"/user/get/professionals":{"get":{"operationId":"UserController_getUsersByCrp","summary":"Get all the data from the professionals","parameters":[],"responses":{"200":{"description":"Returns a JSON with the data from all the professionals","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetUsersByCrpResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["user"]}},"/user/get/patients":{"get":{"operationId":"UserController_getUsersByPatientId","summary":"Get all the data from the patients","parameters":[],"responses":{"200":{"description":"Returns a JSON with the data from all the patients","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetUsersByPatientIdResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["user"],"security":[{"auth":[]}]}},"/user/post":{"post":{"operationId":"UserController_postUser","summary":"Post the user data","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostUserReqDto"}}}},"responses":{"200":{"description":"Returns a JSON with the user data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetUserResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["user"]}},"/user/put":{"put":{"operationId":"UserController_putUser","summary":"Put the user data","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PutUserReqDto"}}}},"responses":{"200":{"description":"Returns a JSON with the user data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetUserResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["user"],"security":[{"auth":[]}]}},"/user/delete":{"delete":{"operationId":"UserController_deleteUser","summary":"Delete the user data","parameters":[],"responses":{"200":{"description":"Returns a JSON with the user status","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeleteUserResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["user"],"security":[{"auth":[]}]}},"/client/get-my-patients-rdps":{"get":{"operationId":"ClientController_getPatientsRDP","summary":"Get the client data","parameters":[{"name":"patientId","required":false,"in":"query","schema":{"type":"string"}},{"name":"startDate","required":false,"in":"query","schema":{"type":"string"}},{"name":"endDate","required":false,"in":"query","schema":{"type":"string"}},{"name":"emotion","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"Returns a JSON with the client data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetClientResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["client"],"security":[{"auth":[]}]}},"/client/get/my-rdps":{"get":{"operationId":"ClientController_getMyRDP","summary":"Get the client data","parameters":[{"name":"startDate","required":false,"in":"query","schema":{"type":"string"}},{"name":"endDate","required":false,"in":"query","schema":{"type":"string"}},{"name":"emotion","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"Returns a JSON with the client data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetClientResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["client"],"security":[{"auth":[]}]}},"/client/post":{"post":{"operationId":"ClientController_postRDP","summary":"Post the client data","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostClientReqDto"}}}},"responses":{"200":{"description":"Returns a JSON with the client data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetClientResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["client"],"security":[{"auth":[]}]}},"/client/put":{"put":{"operationId":"ClientController_putRDP","summary":"Put the user data","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PutClientReqDto"}}}},"responses":{"200":{"description":"Returns a JSON with the user data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetClientResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["client"],"security":[{"auth":[]}]}},"/client/delete/{uuid}":{"delete":{"operationId":"ClientController_deleteRPD","summary":"Delete the user data","parameters":[{"name":"uuid","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Returns a JSON with the user status","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeleteClientResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["client"],"security":[{"auth":[]}]}},"/appointment/get/professional-appointments":{"get":{"operationId":"AppointmentController_getAppointment","summary":"Get the appointment data","parameters":[{"name":"patientId","required":false,"in":"query","schema":{"type":"string"}},{"name":"startDate","required":false,"in":"query","schema":{"type":"string"}},{"name":"endDate","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"Returns a JSON with the appointment data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetAppointmentResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["appointment"],"security":[{"auth":[]}]}},"/appointment/get/my-appointments":{"get":{"operationId":"AppointmentController_getMyAppointments","summary":"Get all the client data appointments","parameters":[],"responses":{"200":{"description":"Returns a JSON with the data from all client data appointments","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetAppointmentResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["appointment"],"security":[{"auth":[]}]}},"/appointment/post":{"post":{"operationId":"AppointmentController_postAppointment","summary":"Post the appointment data","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostAppointmentReqDto"}}}},"responses":{"200":{"description":"Returns a JSON with the appointment data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetAppointmentResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["appointment"],"security":[{"auth":[]}]}},"/appointment/put":{"put":{"operationId":"AppointmentController_putAppointment","summary":"Put the appointment data","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PutAppointmentReqDto"}}}},"responses":{"200":{"description":"Returns a JSON with the appointment data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetAppointmentResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["appointment"],"security":[{"auth":[]}]}},"/appointment/patch/link-appointment/{uuid}":{"patch":{"operationId":"AppointmentController_patchLinkAppointment","summary":"Get all the client data appointments","parameters":[{"name":"uuid","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Returns a JSON with the data from all client data appointments","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetAppointmentResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["appointment"],"security":[{"auth":[]}]}},"/appointment/patch/cancel-appointment/{uuid}":{"patch":{"operationId":"AppointmentController_patchCancelAppointment","summary":"Patch the appointment data to cancel a patientId","parameters":[{"name":"uuid","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Returns a JSON with the appointment status","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetAppointmentResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["appointment"],"security":[{"auth":[]}]}},"/appointment/delete/{uuid}":{"delete":{"operationId":"AppointmentController_deleteAppointment","summary":"Delete the appointment data","parameters":[{"name":"uuid","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Returns a JSON with the appointment status","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeleteAppointmentResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["appointment"],"security":[{"auth":[]}]}},"/professional/get":{"get":{"operationId":"ProfessionalController_getPacientCase","summary":"Get the cases data","parameters":[{"name":"patientId","required":false,"in":"query","schema":{"type":"string"}},{"name":"startDate","required":false,"in":"query","schema":{"type":"string"}},{"name":"endDate","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"Returns a JSON with the professional data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetProfessionalResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["professional"],"security":[{"auth":[]}]}},"/professional/post":{"post":{"operationId":"ProfessionalController_postCase","summary":"Post the case data","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostProfessionalReqDto"}}}},"responses":{"200":{"description":"Returns a JSON with the professional data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetProfessionalResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["professional"],"security":[{"auth":[]}]}},"/professional/put":{"put":{"operationId":"ProfessionalController_putCase","summary":"Put the case data","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PutProfessionalReqDto"}}}},"responses":{"200":{"description":"Returns a JSON with the professional data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetProfessionalResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["professional"],"security":[{"auth":[]}]}},"/professional/delete/{uuid}":{"delete":{"operationId":"ProfessionalController_deleteHistory","summary":"Delete the case history data","parameters":[{"name":"uuid","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Returns a JSON with the professional status","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeleteProfessionalResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["professional"],"security":[{"auth":[]}]}}},"info":{"title":"ingt-api-gateway","description":"Rational Emotive Behaviour Therapy - Dysfunctional Thought Record","version":"1.0.0","contact":{}},"tags":[],"servers":[{"url":"/ingt-api-gateway"}],"components":{"securitySchemes":{"auth":{"scheme":"bearer","bearerFormat":"JWT","type":"http","name":"JWT","in":"header"}},"schemas":{"PostLoginReqDto":{"type":"object","properties":{"email":{"type":"string"},"crp":{"type":"string"},"password":{"type":"string"}},"required":["email","crp","password"]},"PostLoginResDto":{"type":"object","properties":{"jwt":{"type":"string"}},"required":["jwt"]},"ErrorDto":{"type":"object","properties":{"statusCode":{"type":"number"},"message":{"type":"string"}},"required":["statusCode","message"]},"UserDto":{"type":"object","properties":{"name":{"type":"string"},"address":{"type":"string"},"phone":{"type":"string"},"email":{"type":"string"},"birthdate":{"type":"string"},"crp":{"type":"string"},"responsibleCrp":{"type":"string"},"patientId":{"type":"string"}},"required":["name","address","phone","email","birthdate","crp","responsibleCrp","patientId"]},"GetUserResDto":{"type":"object","properties":{"user":{"$ref":"#/components/schemas/UserDto"}},"required":["user"]},"GetUsersByCrpResDto":{"type":"object","properties":{"user":{"type":"array","items":{"type":"string"}}},"required":["user"]},"GetUsersByPatientIdResDto":{"type":"object","properties":{"user":{"type":"array","items":{"type":"string"}}},"required":["user"]},"PostUserReqDto":{"type":"object","properties":{"cpfCnpj":{"type":"string"},"name":{"type":"string"},"address":{"type":"string"},"phone":{"type":"string"},"email":{"type":"string"},"birthdate":{"type":"string"},"crp":{"type":"string"},"responsibleCrp":{"type":"string"},"password":{"type":"string"}},"required":["cpfCnpj","name","address","phone","email","birthdate","crp","responsibleCrp","password"]},"PutUserReqDto":{"type":"object","properties":{"address":{"type":"string"},"phone":{"type":"string"},"password":{"type":"string"}},"required":["address","phone","password"]},"DeleteUserResDto":{"type":"object","properties":{"statusCode":{"type":"number"},"message":{"type":"string"}},"required":["statusCode","message"]},"GetClientResDto":{"type":"object","properties":{"client":{"type":"array","items":{"type":"string"}}},"required":["client"]},"PostClientReqDto":{"type":"object","properties":{"emotion":{"type":"array","items":{"type":"string"}},"physiologicalReaction":{"type":"array","items":{"type":"string"}},"situation":{"type":"string"},"thought":{"type":"string"},"behavior":{"type":"array","items":{"type":"string"}}},"required":["emotion","physiologicalReaction","situation","thought","behavior"]},"PutClientReqDto":{"type":"object","properties":{"uuid":{"type":"string"},"emotion":{"type":"string"},"physiologicalReaction":{"type":"string"},"situation":{"type":"string"},"thought":{"type":"string"},"behavior":{"type":"string"}},"required":["uuid","emotion","physiologicalReaction","situation","thought","behavior"]},"DeleteClientResDto":{"type":"object","properties":{"statusCode":{"type":"number"},"message":{"type":"string"}},"required":["statusCode","message"]},"GetAppointmentResDto":{"type":"object","properties":{"appointment":{"type":"array","items":{"type":"string"}}},"required":["appointment"]},"PostAppointmentReqDto":{"type":"object","properties":{"patientId":{"type":"string"},"date":{"type":"string"},"startTime":{"type":"string"},"endTime":{"type":"string"},"type":{"type":"string"},"location":{"type":"string"}},"required":["patientId","date","startTime","endTime","type","location"]},"PutAppointmentReqDto":{"type":"object","properties":{"uuid":{"type":"string"},"date":{"type":"string"},"startTime":{"type":"string"},"endTime":{"type":"string"},"type":{"type":"string"},"location":{"type":"string"}},"required":["uuid","date","startTime","endTime","type","location"]},"DeleteAppointmentResDto":{"type":"object","properties":{"statusCode":{"type":"number"},"message":{"type":"string"}},"required":["statusCode","message"]},"GetProfessionalResDto":{"type":"object","properties":{"professional":{"type":"array","items":{"type":"string"}}},"required":["professional"]},"PostProfessionalReqDto":{"type":"object","properties":{"patientId":{"type":"string"},"startOfAppointments":{"type":"string"},"responses":{"type":"object"}},"required":["patientId","startOfAppointments","responses"]},"PutProfessionalReqDto":{"type":"object","properties":{"uuid":{"type":"string"},"startOfAppointments":{"type":"string"},"responses":{"type":"object"}},"required":["uuid","startOfAppointments","responses"]},"DeleteProfessionalResDto":{"type":"object","properties":{"statusCode":{"type":"number"},"message":{"type":"string"}},"required":["statusCode","message"]}}}}